[weather][refactor] move CWeatherJob out into its own h/cpp
