+++ xbmc/weather/CMakeLists.txt
@@ -1,7 +1,9 @@
 set(SOURCES GUIWindowWeather.cpp
+            WeatherJob.cpp
             Weather.cpp)
 
 set(HEADERS GUIWindowWeather.h
+            WeatherJob.h
             Weather.h)
 
 core_add_library(weather)
