+++ xbmc/weather/WeatherJob.h
@@ -1,0 +1,74 @@
+#pragma once
+/*
+ *      Copyright (C) 2012-2017 Team Kodi
+ *      http://kodi.tv
+ *
+ *  This Program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2, or (at your option)
+ *  any later version.
+ *
+ *  This Program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with XBMC; see the file COPYING.  If not, see
+ *  <http://www.gnu.org/licenses/>.
+ *
+ */
+
+#include <string>
+#include <map>
+
+#include "Weather.h"
+
+class CWeatherJob : public CJob
+{
+public:
+  explicit CWeatherJob(int location);
+
+  bool DoWork() override;
+
+  const CWeatherInfo &GetInfo() const;
+private:
+  static std::string ConstructPath(std::string in);
+  void LocalizeOverview(std::string &str);
+  void LocalizeOverviewToken(std::string &str);
+  void LoadLocalizedToken();
+  static int ConvertSpeed(int speed);
+
+  void SetFromProperties();
+
+  /*! \brief Formats a celsius temperature into a string based on the users locale
+   \param text the string to format
+   \param temp the temperature (in degrees celsius).
+   */
+  static void FormatTemperature(std::string &text, double temp);
+
+  struct ci_less : std::binary_function<std::string, std::string, bool>
+  {
+    // case-independent (ci) compare_less binary function
+    struct nocase_compare : public std::binary_function<unsigned char,unsigned char,bool>
+    {
+      bool operator() (const unsigned char& c1, const unsigned char& c2) const {
+        return tolower (c1) < tolower (c2);
+      }
+    };
+    bool operator() (const std::string & s1, const std::string & s2) const {
+      return std::lexicographical_compare
+      (s1.begin (), s1.end (),
+       s2.begin (), s2.end (),
+       nocase_compare ());
+    }
+  };
+
+  std::map<std::string, int, ci_less> m_localizedTokens;
+  typedef std::map<std::string, int, ci_less>::const_iterator ilocalizedTokens;
+
+  CWeatherInfo m_info;
+  int m_location;
+
+  static bool m_imagesOkay;
+};
